dictionary and city_to_state modularised
